ai-reports-service/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database (PostgreSQL)
psycopg2-binary==2.9.9

# Authentication
PyJWT==2.8.0
python-jose[cryptography]==3.3.0

# Google Gemini AI
google-generativeai==0.3.2

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6